+++ src/app/modules/resource-table/resource-table.component.ts
@@ -12,16 +12,22 @@ import { HttpClient } from '@angular/common/http';
 import { SelectionModel } from '@angular/cdk/collections';
 import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
 import { ColumnDescription } from '../../types/column.description';
-import { bufferCount } from 'rxjs/operators';
-import { capitalize, escapeStringForRegExp } from '../../shared/utils';
+import { bufferCount, filter } from 'rxjs/operators';
+import { escapeStringForRegExp } from '../../shared/utils';
 import { ColumnDescriptionsService } from '../../shared/column-descriptions/column-descriptions.service';
 import { ColumnValuesService } from '../../shared/column-values/column-values.service';
-import { Subject } from 'rxjs';
+import { Subject, Subscription } from 'rxjs';
 import Resource = fhir.Resource;
 import { TableVirtualScrollDataSource } from 'ng-table-virtual-scroll';
 import { SettingsService } from '../../shared/settings-service/settings.service';
 import { Sort } from '@angular/material/sort';
 import { LiveAnnouncer } from '@angular/cdk/a11y';
+import * as fhirpath from 'fhirpath';
+import * as fhirPathModelR4 from 'fhirpath/fhir-context/r4';
+import {
+  ConnectionStatus,
+  FhirBackendService
+} from '../../shared/fhir-backend/fhir-backend.service';
 
 /**
  * Component for loading table of resources
@@ -47,24 +53,39 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
   isLoading = true;
   loadTime = 0;
   loadedDateTime: number;
+  subscription: Subscription;
+  fhirPathModel: any;
 
   @HostBinding('class.fullscreen') fullscreen = false;
 
   constructor(
+    private fhirBackend: FhirBackendService,
     private http: HttpClient,
     private ngZone: NgZone,
     private columnDescriptionsService: ColumnDescriptionsService,
     private columnValuesService: ColumnValuesService,
     private settings: SettingsService,
     private liveAnnoncer: LiveAnnouncer
-  ) {}
+  ) {
+    this.subscription = fhirBackend.initialized
+      .pipe(filter((status) => status === ConnectionStatus.Ready))
+      .subscribe(() => {
+        this.fhirPathModel = {
+          R4: fhirPathModelR4
+        }[fhirBackend.currentVersion];
+      });
+  }
 
   ngOnInit(): void {}
 
-  ngOnDestroy(): void {}
+  ngOnDestroy(): void {
+    this.subscription.unsubscribe();
+  }
 
   /**
-   * Use columns present in bundle info as default, if empty column descriptions is passed in
+   * Use columns present in bundle info as default, if empty column descriptions is passed in.
+   * We show all columns with data on very first run of the application and when
+   * the user doesn't select any columns to show in the column selection dialog.
    */
   private setColumnsFromBundle(): void {
     // Don't update if no data is available
@@ -78,10 +99,17 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
     const hiddenByDefault = (
       this.settings.get('hideElementsByDefault')?.[this.resourceType] || []
     ).concat(this.settings.get('hideElementsByDefault')?.['*'] || []);
+    // Remove columns without data.
+    // Note: We check all rows, because the first row may
+    // be missing columns that are in the second and subsequent ones.
+    // For example, Patient.interpretation.
+    // I also saw completely blank rows at the beginning of the ResearchStudy
+    // table for dbGap, which was the main reason for checking all rows,
+    // but now these lines are gone.
     this.columnDescriptions = allColumns.filter(
       (x) =>
         hiddenByDefault.indexOf(x.element) === -1 &&
-        this.getCellStrings(this.dataSource.data[0], x).length
+        this.dataSource.data.some((row) => this.getCellStrings(row, x).length)
     );
     // Save column selections of default
     window.localStorage.setItem(
@@ -142,8 +170,8 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
       this.columnDescriptions.forEach((column) => {
         this.filtersForm.addControl(column.element, new FormControl());
       });
-      this.dataSource.filterPredicate = ((data, filter) => {
-        for (const [key, value] of Object.entries(filter)) {
+      this.dataSource.filterPredicate = ((data, filterValues) => {
+        for (const [key, value] of Object.entries(filterValues)) {
           if (value) {
             const columnDescription = this.columnDescriptions.find(
               (c) => c.element === key
@@ -185,6 +213,13 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
   }
 
   /**
+   * Deselect all rows.
+   */
+  clearSelection(): void {
+    this.selectedResources.clear();
+  }
+
+  /**
    * Clear filters on all columns
    */
   clearColumnFilters(): void {
@@ -204,14 +239,12 @@ export class ResourceTableComponent implements OnInit, OnChanges, OnDestroy {
    * @param column - column description
    */
   getCellStrings(row: Resource, column: ColumnDescription): string[] {
-    const fullPath = column.element
-      ? this.resourceType + '.' + column.element
-      : '';
+    const expression = column.expression || column.element.replace('[x]', '');
+    const fullPath = expression ? this.resourceType + '.' + expression : '';
 
     for (const type of column.types) {
-      const element = column.element.replace('[x]', capitalize(type));
       const output = this.columnValuesService.valueToStrings(
-        row[element],
+        fhirpath.evaluate(row, fullPath, {}, this.fhirPathModel),
         type,
         column.isArray,
         fullPath